extract type assertion utilities from member.ts to type-assertion-access.ts
